% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bathymetry.R
\name{calculate_bathymetry}
\alias{calculate_bathymetry}
\title{Calculate bathymetric metrics}
\usage{
calculate_bathymetry(df)
}
\arguments{
\item{df}{a data frame with a "lake", "metric", "variable", and "value"
columns where one of the metrics is "exceedance_level".}
}
\value{
bathymetry, a data frame with "lake", "metric", "variable", and
"value" where the metrics are "volume", "area", "max_depth", and
"mean_depth" and the variables are all the exceedance level
probabilities inputted in df.
}
\description{
Given a data frame with the metrics for a single set of simulations,
calculate the associated volume (m3), area (m2), maximum depth (m), and mean
depth (m) for each exceedance_level (m). Uses CSLSdata::bathymetry and
CSLSdata::lake_raster, only performs calculations on lakes which have
information in all three data sources.
}
