% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lake.R
\name{calculate_lake}
\alias{calculate_lake}
\title{Calculate frequency of lake and flood conditions}
\usage{
calculate_lake(
  df,
  lake_info = data.frame(lake = c("Long", "Plainfield"), min_elev_m = c(333.8985,
    333.1118), max_elev_m = c(335.536, 335.0016))
)
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{lake_info}{a data frame with the "min_elev_m" and "max_elev_m" at which
each "lake" becomes a lake/good for paddleboating and floods
back vegetation.}
}
\value{
a data frame with the number of months and percent of time with
lake, good paddleboating, and flooded vegetation conditions at each
lake included in "lake_info".
}
\description{
Given a data frame with columns for "lake" and "level" as well as additional
information about the elevation at which lake becomes a lake and suitable for
paddleboating and the elevation at which the lake floods back floating leaf
plants, evaluate the frequency of lake conditions, good paddleboating
conditions during warm months (April thorugh September), and flooded
conditions at all lakes with lake information.
}
