% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_motorboat.R
\name{calculate_motorboat}
\alias{calculate_motorboat}
\title{Calculate frequency of legal motorboat conditions}
\usage{
calculate_motorboat(
  df,
  boat_info = data.frame(lake = "Pleasant", elev_m = 292.6422)
)
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{boat_info}{a data frame with the "elev_m" at which each "lake"
switches from legal to not legal for motorboating.}
}
\value{
a data frame with the number of months and percent of time with
mixed conditions at each lake included in "boat_info".
}
\description{
Given a data frame with columns for "lake" and "level" as well as additional
information about the elevation at which lakes are suitable for motorboating
(lake area > 62.5 acres), evaluate the frequency of legal motorboating
conditions at all lakes with motorboat information.
}
