% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_paddleboat.R
\name{calculate_paddleboat}
\alias{calculate_paddleboat}
\title{Calculate frequency of good paddleboat conditions}
\usage{
calculate_paddleboat(
  df,
  paddle_info = data.frame(lake = "Long", elev_m = 334.29)
)
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{paddle_info}{a data frame with the "elev_m" at which each "lake"
becomes good for paddleboating.}
}
\value{
a data frame with the number of months and percent of time with
good paddleboating conditions at each lake included in "paddle_info".
}
\description{
Given a data frame with columns for "lake" and "level" as well as additional
information about the elevation at which lake becomes suitable for
paddleboating, evaluate the frequency of good paddleboating conditions during
warm months (April thorugh September) at all lakes with paddleboat
information.
}
