% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_plant_area.R
\name{calculate_plant_area}
\alias{calculate_plant_area}
\title{Calculate area of each vegetation class corresponding to lake elevations}
\usage{
calculate_plant_area(df)
}
\arguments{
\item{df}{a data frame with a "lake", "metric", "variable", and "value"
columns where one of the metrics is "exceedance_level".}
}
\value{
plant_areas, a data frame with "lake", "metric", "variable", and
"value" where the metrics are plant community types and the variables
are all the exceedance level probabilities inputted in df, and the
values are the corresponding plant areas in m^2.
}
\description{
Given a metrics data frame with values for exceedance_level, use the
information in CSLSdata::bathymetry to calculate the area of each plant
community at each exceedance level.
}
