% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stratification.R
\name{calculate_stratification}
\alias{calculate_stratification}
\title{Calculate frequency of mixed (unstratified) conditions}
\usage{
calculate_stratification(
  df,
  strat_info = data.frame(lake = "Pleasant", elev_m = 297.67)
)
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{strat_info}{a data frame with the "elev_m" at which each "lake"
switches from stratified to mixed.}
}
\value{
a data frame with the number of months and percent of time with
mixed conditions at each lake included in "strat_info".
}
\description{
Given a data frame with columns for "lake" and "level" as well as additional
information about the elevation at which lakes stratify, evaluate the
frequency of mixed conditions at all lakes with stratification information.
}
