% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_substrate_hardness.R
\name{calculate_substrate_hardness}
\alias{calculate_substrate_hardness}
\title{Calculate centrarchid substrate area at given Pleasant Lake elevation}
\usage{
calculate_substrate_hardness(df)
}
\arguments{
\item{df}{a data frame with a "lake", "metric", "variable", and "value"
columns where one of the metrics is "exceedance_level".}
}
\value{
substrate, a data frame with "lake", "metric", "variable", and
"value" where the metric is "centrarchid_substrate", the variables
are all the exceedance level probabilities inputted in df, and the
value is the mean substrate hardness at generalized centrarchid
spawning depth.
}
\description{
Given a data frame with the exceedance level metrics for a single set of
simulations at Pleasant Lake, calculate the associated substrate area for
centrarchid fish. Uses CSLSdata::fish_substrate, but converts areas from
acres to m^2 to be consistent with other area metrics.
}
