% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_turtle.R
\name{calculate_turtle}
\alias{calculate_turtle}
\title{Calculate frequency of connection to Turtle Bay}
\usage{
calculate_turtle(df, turtle_elev_m = 298.7, turtle_clearance_ft = 1)
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{turtle_elev_m}{elevation of turtle bay inlet by eye examination of
bathymetry raster (m).}

\item{turtle_clearance_ft}{minimum clearnace needed for good connection to
turtle bay.}
}
\value{
a data frame with the number of months and percent of time with
connection to Turtle Bay during all months and during warm season
(March through August) months.
}
\description{
Given a data frame with columns for "lake", "level", "month", and "year" as
well as additional information about the elevation of Turtle Bay
("turtle_elev_m") and minimum required clearance for good connection
("turtle_clearance_ft"), calculates frequency of connection at all times and
during the growing season (March through August).
}
