% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stratification.R
\name{check_stratification}
\alias{check_stratification}
\title{Calculate frequency of mixed (unstratified) conditions}
\usage{
check_stratification(
  this_rule,
  this_hydro,
  metric_uncertainty,
  strat_info = data.frame(lake = "Pleasant", elev_m = 297.67)
)
}
\arguments{
\item{this_rule}{data frame with information on "percent", "difference",
"significant_if" (i.e., "higher" or "lower") for a single
indicator.}

\item{this_hydro}{data frame with the hydrologic metrics (solute_budget) to
evaluate.}

\item{metric_uncertainty}{data frame with lake, metric, variable, and
allowable "difference" due to uncertainty in the
metric. Currently evaluated as the standard
deviation in the "no irrigation" scenarios of
the metric.}

\item{strat_info}{a data frame with the "elev_m" at which each "lake"
switches from stratified to mixed.}
}
\value{
this_hydro, a data frame noting thresholds and impact.
}
\description{
Given a data frame with columns for "lake" and "level" as well as additional
information about the elevation at which lakes stratify, evaluate the
frequency of mixed conditions at all lakes with stratification information.
}
