% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_scenarios.R
\name{compare_scenarios}
\alias{compare_scenarios}
\title{Compare scenarios}
\usage{
compare_scenarios(
  df1,
  df2,
  metric_uncertainty,
  rules = CSLSscenarios::ecological_rules,
  bathymetry = CSLSdata::bathymetry
)
}
\arguments{
\item{df1}{data frame with baseline hydrologic metrics}

\item{df2}{data frame with hydrologic metrics of scenario being evaluated for
significant impact.}

\item{metric_uncertainty}{data frame with lake, metric, variable, and
allowable "difference" due to uncertainty in the
metric. Currently evaluated as the standard
deviation in the "no irrigation" scenarios of
the metric.}

\item{rules}{data frame with ecological rules for ecological indicators
related to hydrologic metrics.}

\item{bathymetry}{data frame with bathymetric relationships with parameters like
lake area, lake volume, plant area, etc.}
}
\value{
comparison, a data frame with the following columns:
\item{lake}{name of lake}
\item{hydrology}{type of hydrologic metric (e.g., magnitude)}
\item{metric}{hydrologic metric (e.g., exceedance_level)}
\item{variable}{type of hydrologic metric (e.g. "90" for 90th percentile
exceedance level)}
\item{category}{category of ecological indicator (e.g., plants, fish)}
\item{indicator}{ecological indicator (e.g., volume_habitat)}
\item{impacted}{logical, TRUE if this ecological indicator is impacted from
baseline under this scenario}
\item{significant_if}{notes whether scenario is significant if scenario
values are "higher" or "lower" than threshold values}
\item{value1}{base value of hydrologic metric}
\item{threshold}{threshold value of hydrologic metric}
\item{value2}{scenario value of hydrologic metric}
\item{threshold_diff}{difference between threshold and base value of
hydrologic metric}
\item{diff}{difference between scenario and base value of hydrologic metric}
\item{bathy_significant_if}{notes whether scenario is significant if scenario
child/bathymetric values are "higher" or "lower"
than child/bathymetric threshold values}
\item{bathy1}{base value of child/bathymetric metric (if applicable)}
\item{bathy_threshold}{threshold value of child/bathymetric metric (if
applicable)}
\item{bathy2}{scenario value of child/bathymetric metric (if applicable)}
\item{bathy_threshold_diff}{difference between bathy_threshold and bathy1}
\item{bathy_diff}{difference between bathy2 and bathy1}
}
\description{
This function compares two time series and evaluates the second series for
ecologically significant differences from the first.
}
