% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_solute_median.R
\name{evaluate_solute_median}
\alias{evaluate_solute_median}
\title{Evaluate median solute concentration}
\usage{
evaluate_solute_median(this_hydro, this_rule, metric_uncertainty)
}
\arguments{
\item{this_hydro}{data frame with the hydrologic metrics (solute_budget) to
evaluate.}

\item{this_rule}{data frame with information on "percent", "difference",
"significant_if" (i.e., "higher" or "lower") for a single
indicator.}

\item{metric_uncertainty}{data frame with lake, metric, variable, and
allowable "difference" due to uncertainty in the
metric. Currently evaluated as the standard
deviation in the "no irrigation" scenarios of
the metric.}
}
\value{
this_hydro, a data frame noting thresholds and impact.
}
\description{
Given metrics related to solute concentration, evaluate whether new median
concentration is lower thant the 90\% concentration or higher than the 10\%
concentration.
}
