% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bounds.R
\name{select_bounds}
\alias{select_bounds}
\title{Select upper/lower bounds for comparisons}
\usage{
select_bounds(df, base_scenario = "no_irr", compare_scenario = "irr")
}
\arguments{
\item{df}{a data frame with MODFLOW metrics}

\item{base_scenario}{name of base scenario, defaults to "no_irr"}

\item{compare_scenario}{name of scenario to evaluate for impacts relative to
the base scenario, defaults to "irr".}
}
\value{
use_sims, a data frame with the following columns:
\item{sim}{simulations numbers to use}
\item{sim_type}{notes whether this simulation represents a "conservative" or
"permissive" determination of impact}
}
\description{
Selects a "conservative" and "permissive" MODFLOW simulation for evaluation
of impacts.
}
