% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_rank_dist.R
\docType{data}
\name{well_rank_dist}
\alias{well_rank_dist}
\title{Dataset: Distance of wells from centroild of Long/Pleasant lakes}
\format{
A data frame with the following columns.
\describe{
\item{lake}{name of lake}
\item{sim}{well rank aka simulation number. smaller = closer to lake.}
\item{dist_m}{distance from well to centroid of Long Lake (for Plainfield &
Long) or Pleasant Lake (for Pleasant), (m)}
\item{dist_mi}{distance from well to centroid of Long Lake (for Plainfield &
Long) or Pleasant Lake (for Pleasant), (miles)}
}
}
\usage{
data(well_rank_dist)
}
\description{
Dataset: Distance of wells from centroild of Long/Pleasant lakes
}
\keyword{datasets}
