% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_bathymetry.R
\name{extrapolate_bathymetry}
\alias{extrapolate_bathymetry}
\title{Extrapolate relationship between elevation and bathymetric metrics}
\usage{
extrapolate_bathymetry(
  this_hydro,
  this_rule,
  bathymetry,
  bathy_metric,
  metric_uncertainty
)
}
\arguments{
\item{this_hydro}{data frame with the hydrologic metrics (exceedance levels) to
evaluate.}

\item{this_rule}{data frame with the ecological rules for given parameter to
evaluate.}

\item{bathymetry}{data frame with relationships between lake elevation and
other parameters (e.g., lake area, lake volume, plant area,
substrate area)}

\item{bathy_metric}{name of column in bathymetry to evaluate}

\item{metric_uncertainty}{data frame with lake, metric, variable, and
allowable "difference" due to uncertainty in the
metric. Currently evaluated as the standard
deviation in the "no irrigation" scenarios of
the metric.}
}
\value{
impact_evaluation, a data frame with the following columns:
\item{lake}{name of lake, character}
\item{metric}{hydrologic metric, i.e. "exceedance_level"}
\item{variable}{hydrologic metric variable, i.e. 10, 25, 50, 75, or 90}
\item{value1}{baseline value for hydrologic metric}
\item{threshold}{threshold value for hydrologic metric for impact}
\item{value2}{scenario value for hydrologic metric}
\item{impacted}{logical, indicates whether lake is impacted relative to this
ecological indicator under this scenario (TRUE) or not
(FALSE)}
}
\description{
Assumes ecological rule is a percent change (increase or decrease) from
baseline value (e.g., 10\% decrease), not a numerical difference (e.g., 2
fewer times).
}
\details{
Because so many bathymetric relationships are non-monotonic, had to come up
with a hacky way of finding correct elevation match for given threshold
child bathy parameter. From check on monotonic child parameters, should yield
accurate threshold elevations to within 1 cm.
}
